chore: centralize run_id sanitizer and harden CLI errors
